move common.rb to a proper path
