+++ lib/puppet/provider/cert_file/posix.rb
@@ -1,7 +1,7 @@
 # frozen_string_literal: true
 
 require 'pp'
-require 'common'
+require 'puppet_x/common'
 
 Puppet::Type.type(:cert_file).provide :posix do
   confine feature: :posix
