+++ lib/puppet_x/common.rb
